requirements.txt
python-dotenv 
typing-extensions 
langgraph 
langchain-core 
langchain-openai 
langchain-community 
openai 
ipython 
pydantic 
langgraph-cli[inmem]
langgraph-prebuilt